Zaliczenie Java

+++ src/main/java/pl/pjatk/s20685bank/Model/Account.java
@@ -10,7 +10,7 @@ import lombok.With;
 @AllArgsConstructor
 public class Account {
 
-    private Long id;
+    private long id;
     private String name;
-    private Long amount;
+    private double amount;
 }
